--- a/code/aind_auto_train/curriculums/uncoupled_baiting_2p0.py
+++ b/code/aind_auto_train/curriculums/uncoupled_baiting_2p0.py
@@ -78,8 +78,8 @@ paras_stage_1_warmup = DynamicForagingParas(
     # Auto water
     AutoReward=True,
     AutoWaterType=AutoWaterMode.NATURAL,
-    Unrewarded=5,
-    Ignored=5,
+    Unrewarded=3,
+    Ignored=3,
     Multiplier=0.5,
 
     # Auto block
@@ -134,6 +134,9 @@ paras_stage_1 = DynamicForagingParas(
             # -- Essentials --
             # Turn off Warmup from now on
             warmup='off',
+            
+            Unrewarded=5,
+            Ignored=5,      
             
             # Decrease water size to 2.0 from now on
             RightValue_volume=2.0,