… code/aind_auto_train/curriculums/uncoupled_no_baiting_2p0.py