### Trying to combat the motivation issue
- Decreased `reward size`
  Old (uncoupled baiting):
![image](https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/7c1d0562-2352-429e-9644-7f5cb504a33c)
  New (uncoupled baiting):
  ![image](https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/797793ed-535b-4d73-b5fe-ded2b88a3113)

### Decrease autowater more rapidly
- resolves https://github.com/AllenNeuralDynamics/aind-behavior-blog/discussions/290#discussioncomment-8762278 @ellahiltonvano 
 Old:
    <img src="https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/25bc7b84-6035-4d58-bb1a-5567a7017db6" width=600>
 New: 
   <img src="https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/57b73506-9efd-4e7e-9be7-21746485b5ad" width=600>

### ResponseTime and RewardConsumeTime
- resolves https://github.com/AllenNeuralDynamics/aind-behavior-blog/issues/309 @XX-Yin 
- Decreased `ResponseTime` in the final stage
- Use lower `RewardConsumeTime` at earlier stage to speed up trials but increase it to a normal 3.0 s in the final stage
   Old: 
    ![image](https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/5a7aa25b-783e-429d-9cb8-ad52a96da120)
    New:
    ![image](https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/5d1cdda2-b322-446a-8e1c-936ed288c1fb)

### Increased progress criteria a bit
- Because (1) we decreased the reward size, and (2) previous v1.0 mice are doing good in terms of foraging efficiency

| Old | New | 
|-----|-----|
|<img src="https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/f3770186-35cd-4966-9d82-90fdf3feea90" width=350> |<img src="https://github.com/AllenNeuralDynamics/aind-foraging-behavior-bonsai-automatic-training/assets/24734299/ccfa4bcc-f9a3-483c-b022-05b58d8f9c39" width=350> |